Add Animation Event after Level Complete to Load next Scene
We attach a Script to Level Complete Panel, as it is one being animated. The Level complete panel becomes Active when Player touches EndGame Object. First the Animation plays. keep the Animation Event  after it. In the Inspector you can choose the function to run. Since the script is attached to LevelCompletePanel, you can add the Load Next function. You need to create prefabs of player and other essential elements so that they can be modified uniformly across levels when needed. Then Duplicate lvl1 Add the new scenes in the Build settings.