Add button to EndCredits Scene, Duplicate to make Title Scene
Create the QuitGame script and attach it to Credits Panel. Then Add On Click Event on Button. Drag and drop the Panel Object into the slot. Choose the Quit Game function from the list.
In the Opening Credits Create a new PlayGame Script and attach it to the panel. Delete the existing script and repeat the process. The function makes use of SceneManagement to load the level 1 scene